AgenticDataConnect/requirements.txt
duckdb
pandas
pyyaml
rapidfuzz
pydantic
google-generativeai
fastapi
uvicorn
httpx
sift-stack-py
